+++ hspec-core/src/Test/Hspec/Core/Formatters/V2.hs
@@ -30,10 +30,12 @@ module Test.Hspec.Core.Formatters.V2 (
 -- ** Accessing the runner state
 , getSuccessCount
 , getPendingCount
+, getFailureCount
 , getFailCount
 , getTotalCount
 
 , FailureRecord (..)
+, getFailureMessages
 , getFailMessages
 , usedSeed
 
@@ -52,6 +54,7 @@ module Test.Hspec.Core.Formatters.V2 (
 , withInfoColor
 , withSuccessColor
 , withPendingColor
+, withFailureColor
 , withFailColor
 
 , useDiff
@@ -88,10 +91,12 @@ import Test.Hspec.Core.Formatters.Monad (
 
   , getSuccessCount
   , getPendingCount
+  , getFailureCount
   , getFailCount
   , getTotalCount
 
   , FailureRecord (..)
+  , getFailureMessages
   , getFailMessages
   , usedSeed
 
@@ -106,6 +111,7 @@ import Test.Hspec.Core.Formatters.Monad (
   , withInfoColor
   , withSuccessColor
   , withPendingColor
+  , withFailureColor
   , withFailColor
 
   , useDiff
@@ -139,7 +145,7 @@ checks = specdoc {
     uncurry (writeResult nesting requirement (itemDuration item) (itemInfo item)) $ case itemResult item of
       Success {} -> (withSuccessColor, "✔")
       Pending {} -> (withPendingColor, "‐")
-      Failure {} -> (withFailColor, "✘")
+      Failure {} -> (withFailureColor, "✘")
     case itemResult item of
       Success {} -> return ()
       Failure {} -> return ()
@@ -190,8 +196,8 @@ specdoc = silent {
       Pending _ reason -> withPendingColor $ do
         writeResult nesting requirement duration info
         writeLine $ indentationFor ("" : nesting) ++ "# PENDING: " ++ fromMaybe "No reason given" reason
-      Failure {} -> withFailColor $ do
-        n <- getFailCount
+      Failure {} -> withFailureColor $ do
+        n <- getFailureCount
         writeResult nesting (requirement ++ " FAILED [" ++ show n ++ "]") duration info
 
 , formatterDone = defaultFailedFormatter >> defaultFooter
@@ -220,7 +226,7 @@ progress = failed_examples {
   formatterItemDone = \ _ item -> case itemResult item of
     Success{} -> withSuccessColor $ write "."
     Pending{} -> withPendingColor $ write "."
-    Failure{} -> withFailColor $ write "F"
+    Failure{} -> withFailureColor $ write "F"
 }
 
 failed_examples :: Formatter
@@ -232,7 +238,7 @@ defaultFailedFormatter :: FormatM ()
 defaultFailedFormatter = do
   writeLine ""
 
-  failures <- getFailMessages
+  failures <- getFailureMessages
 
   unless (null failures) $ do
     writeLine "Failures:"
@@ -253,7 +259,7 @@ defaultFailedFormatter = do
       writeLine (formatRequirement path)
       case reason of
         NoReason -> return ()
-        Reason err -> withFailColor $ indent err
+        Reason err -> withFailureColor $ indent err
         ExpectedButGot preface expected actual -> do
           mapM_ indent preface
 
@@ -276,21 +282,21 @@ defaultFailedFormatter = do
               (xs, _ : ys) -> output (xs ++ "\n") >> write (indentation ++ "          ") >> indented output ys
 
             writeDiff chunks extra missing = do
-              withFailColor $ write (indentation ++ "expected: ")
+              withFailureColor $ write (indentation ++ "expected: ")
               forM_ chunks $ \ chunk -> case chunk of
                 Both a _ -> indented write a
                 First a -> indented extra a
                 Second _ -> return ()
               writeLine ""
 
-              withFailColor $ write (indentation ++ " but got: ")
+              withFailureColor $ write (indentation ++ " but got: ")
               forM_ chunks $ \ chunk -> case chunk of
                 Both a _ -> indented write a
                 First _ -> return ()
                 Second a -> indented missing a
               writeLine ""
 
-        Error _ e -> withFailColor . indent $ (("uncaught exception: " ++) . formatException) e
+        Error _ e -> withFailureColor . indent $ (("uncaught exception: " ++) . formatException) e
 
       writeLine ""
       writeLine ("  To rerun use: --match " ++ show (joinPath path))
@@ -307,7 +313,7 @@ defaultFooter = do
     <$> (printf "Finished in %1.4f seconds" <$> getRealTime)
     <*> (maybe "" (printf ", used %1.4f seconds of CPU time") <$> getCPUTime)
 
-  fails   <- getFailCount
+  fails   <- getFailureCount
   pending <- getPendingCount
   total   <- getTotalCount
 
@@ -316,7 +322,7 @@ defaultFooter = do
          pluralize total   "example"
       ++ ", " ++ pluralize fails "failure"
       ++ if pending == 0 then "" else ", " ++ show pending ++ " pending"
-    c | fails /= 0   = withFailColor
+    c | fails /= 0   = withFailureColor
       | pending /= 0 = withPendingColor
       | otherwise    = withSuccessColor
   c $ writeLine output