+++ CHANGES.markdown
@@ -2,6 +2,9 @@
   - Better support for `--color=auto` on Windows
   - Add support for spec hooks to `hspec-discover` (via `SpecHooks.hs`)
   - Propagate exceptions from `Test.Hspec.Core.Format.monadic`
+  - Formatters API: Add `getFailureCount`, `getFailureMessages`,
+    `withFailureColor` and deprecate `getFailCount`, `getFailMessages`,
+    `withFailColor`
 
 ## Changes in 2.8.2
   - Add `mapSubject` and `ignoreSubject`